+++ Sources/RedBlackTreeModule/RedBlackTreeDictionary.swift
@@ -86,8 +86,8 @@ extension RedBlackTreeDictionary: ___RedBlackTreeUnique {}
 extension RedBlackTreeDictionary: ___RedBlackTreeMerge {}
 extension RedBlackTreeDictionary: ___RedBlackTreeSequenceBase {}
 extension RedBlackTreeDictionary: KeyValueComparer {}
-extension RedBlackTreeDictionary: ElementComparable where Value: Comparable { }
-extension RedBlackTreeDictionary: ElementEqutable where Value: Equatable { }
+extension RedBlackTreeDictionary: ElementComparable where Value: Comparable {}
+extension RedBlackTreeDictionary: ElementEqutable where Value: Equatable {}
 
 // MARK: - Creating a Dictionay
 
@@ -364,17 +364,17 @@ extension RedBlackTreeDictionary {
   @inline(__always)
   public subscript<R>(bounds: R) -> SubSequence where R: RangeExpression, R.Bound == Index {
     let bounds: Range<Index> = bounds.relative(to: self)
-    
+
     __tree_.___ensureValidRange(
       begin: bounds.lowerBound.rawValue,
       end: bounds.upperBound.rawValue)
-    
+
     return .init(
       tree: __tree_,
       start: bounds.lowerBound.rawValue,
       end: bounds.upperBound.rawValue)
   }
-  
+
   /// - Warning: This subscript trades safety for performance. Using an invalid index results in undefined behavior.
   /// - Complexity: O(1)
   @inlinable
@@ -385,12 +385,13 @@ extension RedBlackTreeDictionary {
       start: bounds.lowerBound.rawValue,
       end: bounds.upperBound.rawValue)
   }
-  
+
   /// - Warning: This subscript trades safety for performance. Using an invalid index results in undefined behavior.
   /// - Complexity: O(1)
   @inlinable
   @inline(__always)
-  public subscript<R>(unchecked bounds: R) -> SubSequence where R: RangeExpression, R.Bound == Index {
+  public subscript<R>(unchecked bounds: R) -> SubSequence where R: RangeExpression, R.Bound == Index
+  {
     let bounds: Range<Index> = bounds.relative(to: self)
     return .init(
       tree: __tree_,
@@ -422,8 +423,8 @@ extension RedBlackTreeDictionary {
     inserted: Bool, memberAfterInsert: Element
   ) {
     _ensureUniqueAndCapacity()
-    _ = __tree_.__insert_unique(newMember)
-    return (true, newMember)
+    let (__r, __inserted) = __tree_.__insert_unique(newMember)
+    return (__inserted, __inserted ? newMember : __tree_[__r])
   }
 }
 
@@ -483,7 +484,7 @@ extension RedBlackTreeDictionary {
     _ensureUnique()
     try ___merge_unique(other, uniquingKeysWith: combine)
   }
-  
+
   /// 辞書に `other` の要素をマージします。
   /// キーが重複したときは `combine` の戻り値を採用します。
   ///
@@ -524,7 +525,7 @@ extension RedBlackTreeDictionary {
     try result.merge(other, uniquingKeysWith: combine)
     return result
   }
-  
+
   /// `self` と `other` をマージした新しい辞書を返します。
   ///
   /// - Complexity: O(*n* log(*m + n*)), where *n* is the length of `other`
@@ -822,9 +823,7 @@ extension RedBlackTreeDictionary {
 // MARK: - Collection
 // MARK: - BidirectionalCollection
 
-extension RedBlackTreeDictionary: Sequence, Collection, BidirectionalCollection {}
-
-extension RedBlackTreeDictionary {
+extension RedBlackTreeDictionary: Sequence, Collection, BidirectionalCollection {
 
   /// - Complexity: O(1)
   @inlinable
@@ -838,7 +837,8 @@ extension RedBlackTreeDictionary {
   public func forEach(_ body: (_Value) throws -> Void) rethrows {
     try _forEach(body)
   }
-  
+
+  /// 特殊なforEach
   @inlinable
   @inline(__always)
   public func forEach(_ body: (Index, _Value) throws -> Void) rethrows {
@@ -851,78 +851,78 @@ extension RedBlackTreeDictionary {
   public __consuming func sorted() -> Tree.ElementIterator {
     .init(tree: __tree_, start: __tree_.__begin_node, end: __tree_.__end_node())
   }
-  
+
   /// - Complexity: O(1)
   @inlinable
   @inline(__always)
   public var startIndex: Index { _startIndex }
-  
+
   /// - Complexity: O(1)
   @inlinable
   @inline(__always)
   public var endIndex: Index { _endIndex }
-  
+
   /// - Complexity: O(log *n*)
   @inlinable
   //  @inline(__always)
   public func distance(from start: Index, to end: Index) -> Int {
     _distance(from: start, to: end)
   }
-  
+
   /// - Complexity: O(1)
   @inlinable
   @inline(__always)
   public func index(after i: Index) -> Index {
     _index(after: i)
   }
-  
+
   /// - Complexity: O(1)
   @inlinable
   @inline(__always)
   public func formIndex(after i: inout Index) {
     _formIndex(after: &i)
   }
-  
+
   /// - Complexity: O(1)
   @inlinable
   @inline(__always)
   public func index(before i: Index) -> Index {
     _index(before: i)
   }
-  
+
   /// - Complexity: O(1)
   @inlinable
   @inline(__always)
   public func formIndex(before i: inout Index) {
     _formIndex(before: &i)
   }
-  
+
   /// - Complexity: O(*d*)
   @inlinable
   //  @inline(__always)
   public func index(_ i: Index, offsetBy distance: Int) -> Index {
     _index(i, offsetBy: distance)
   }
-  
+
   /// - Complexity: O(*d*)
   @inlinable
   //  @inline(__always)
   public func formIndex(_ i: inout Index, offsetBy distance: Int) {
     _formIndex(&i, offsetBy: distance)
   }
-  
+
   /// - Complexity: O(*d*)
   @inlinable
   //  @inline(__always)
   public func index(_ i: Index, offsetBy distance: Int, limitedBy limit: Index) -> Index? {
     _index(i, offsetBy: distance, limitedBy: limit)
   }
-  
+
   /// - Complexity: O(*d*)
   @inlinable
   //  @inline(__always)
   public func formIndex(_ i: inout Index, offsetBy distance: Int, limitedBy limit: Index)
-  -> Bool
+    -> Bool
   {
     _formIndex(&i, offsetBy: distance, limitedBy: limit)
   }
@@ -932,7 +932,7 @@ extension RedBlackTreeDictionary {
   public subscript(position: Index) -> _Value {
     @inline(__always) get { self[_checked: position] }
     // コンパイラがクラッシュする
-//    @inline(__always) _read { yield self[_checked: position] }
+    //    @inline(__always) _read { yield self[_checked: position] }
   }
 
   /// - Warning: This subscript trades safety for performance. Using an invalid index results in undefined behavior.
@@ -950,32 +950,31 @@ extension RedBlackTreeDictionary {
   public func isValid(index: Index) -> Bool {
     _isValid(index: index)
   }
-  
+
   /// RangeExpressionがsubscriptやremoveで利用可能か判別します
   ///
   /// - Complexity: O(1)
   @inlinable
   @inline(__always)
-  public func isValid<R: RangeExpression>(
-    _ bounds: R
-  ) -> Bool where R.Bound == Index {
+  public func isValid<R: RangeExpression>(_ bounds: R) -> Bool
+  where R.Bound == Index {
     _isValid(bounds)
   }
-
+  
   /// - Complexity: O(1)
   @inlinable
   @inline(__always)
   public __consuming func reversed() -> Tree.ReversedElementIterator {
     _reversed()
   }
-  
+
   /// - Complexity: O(1)
   @inlinable
   @inline(__always)
   public var indices: Indices {
     _indices
   }
-  
+
   /// - Complexity: O(*m*), where *m* is the lesser of the length of the
   ///   sequence and the length of `other`.
   @inlinable
@@ -985,7 +984,7 @@ extension RedBlackTreeDictionary {
   ) rethrows -> Bool where OtherSequence: Sequence {
     try _elementsEqual(other, by: areEquivalent)
   }
-  
+
   /// - Complexity: O(*m*), where *m* is the lesser of the length of the
   ///   sequence and the length of `other`.
   @inlinable
@@ -1017,7 +1016,7 @@ extension RedBlackTreeDictionary {
 // MARK: - SubSequence
 
 extension RedBlackTreeDictionary {
-  
+
   public typealias SubSequence = RedBlackTreeSlice<Self>
 }
 
@@ -1214,13 +1213,7 @@ extension RedBlackTreeDictionary where Value: Comparable {
 // MARK: - Sendable
 
 #if swift(>=5.5)
+  // TODO: 競プロ用としてはSendableでいいが、一般用としてはSendableが適切かどうか検証が必要
   extension RedBlackTreeDictionary: @unchecked Sendable
   where Element: Sendable {}
-#endif
-
-#if false
-#if swift(>=5.5)
-  extension RedBlackTreeDictionary.SubSequence: @unchecked Sendable
-  where Key: Sendable, Value: Sendable {}
-#endif
 #endif