+++ Sources/RedBlackTreeModule/_Tree/Protocols/Tree_KeyValue.swift
@@ -29,4 +29,3 @@ public protocol Tree_KeyValue {
   static func value_comp(_ lhs: Key,_ rhs: Key) -> Bool
   static func value_equiv(_ :Key,_ :Key) -> Bool
 }
-