+++ Tests/RedBlackTreeTests/multimap/MultiMapTests.swift
@@ -1045,5 +1045,34 @@ import XCTest
         XCTAssertEqual(a + [], [(0, 10),(0, 20),(1, 30),(1, 40)].map{ Pair($0) })
       }
     }
+    
+    func testIsValidRangeSmoke() throws {
+      let a = RedBlackTreeMultiMap<Int,Int>(naive: [0,1,2,3,4,5].map{ Pair($0,$0) })
+      XCTAssertTrue(a.isValid(a.lowerBound(2)..<a.upperBound(4)))
+    }
+    
+    func testSortedReversed() throws {
+      let source = [0,1,2,3,4,5].map { Pair($0,$0 * 10) }
+      let a = RedBlackTreeMultiMap<Int,Int>(multiKeysWithValues: source)
+      XCTAssertEqual(a.sorted() + [], source)
+      XCTAssertEqual(a.reversed() + [], source.reversed())
+    }
+    
+    func testForEach_enumeration() throws {
+      let source = [0,1,2,3,4,5].map { Pair($0,$0 * 10) }
+      let a = RedBlackTreeMultiMap<Int,Int>(multiKeysWithValues: source)
+      var p: RedBlackTreeMultiMap<Int,Int>.Index? = a.startIndex
+      a.forEach { i, v in
+        XCTAssertEqual(i, p)
+        XCTAssertEqual(a[p!], v)
+        p = p?.next
+      }
+    }
+    
+    func testInitNaive_with_Sequence() throws {
+      let source = [0,1,2,3,4,5].map { Pair($0,$0 * 10) }
+      let a = RedBlackTreeMultiMap<Int,Int>(naive: AnySequence(source))
+      XCTAssertEqual(a.sorted() + [], source)
+    }
   }
 #endif