+++ Tests/RedBlackTreeTests/multiset/MultisetTests.swift
@@ -1429,5 +1429,34 @@ final class MultisetTests: XCTestCase {
       }
     }
   }
+  
+  func testIsValidRangeSmoke() throws {
+    let a = RedBlackTreeMultiSet<Int>(naive: [0,1,2,3,4,5])
+    XCTAssertTrue(a.isValid(a.lowerBound(2)..<a.upperBound(4)))
+  }
+  
+  func testSortedReversed() throws {
+    let source = [0,1,2,3,4,5]
+    let a = RedBlackTreeMultiSet<Int>(naive: source)
+    XCTAssertEqual(a.sorted() + [], source)
+    XCTAssertEqual(a.reversed() + [], source.reversed())
+  }
+  
+  func testForEach_enumeration() throws {
+    let source = [0,1,2,3,4,5]
+    let a = RedBlackTreeMultiSet<Int>(naive: source)
+    var p: RedBlackTreeMultiSet<Int>.Index? = a.startIndex
+    a.forEach { i, v in
+      XCTAssertEqual(i, p)
+      XCTAssertEqual(a[p!], v)
+      p = p?.next
+    }
+  }
+  
+  func testInitNaive_with_Sequence() throws {
+    let source = [0,1,2,3,4,5]
+    let a = RedBlackTreeMultiSet<Int>(naive: AnySequence(source))
+    XCTAssertEqual(a.sorted() + [], source)
+  }
 }