+++ Tests/RedBlackTreeTests/tree/NodeFlagTests.swift
@@ -43,6 +43,14 @@ final class NodeFlagTests0_10_20: TreeFixture0_10_20 {
       XCTAssertEqual(___ptr_comp_multi(b, a), ___ptr_comp_bitmap(b, a))
     }
   }
+  
+  func testNodeFlag128() {
+    XCTAssertEqual(___ptr_bitmap_128(__root()), 1 << (UInt128.bitWidth - 1))
+    for (a,b) in (0...2).permutations(ofCount: 2).map({ ($0[0],$0[1])}) {
+      XCTAssertEqual(___ptr_comp_multi(a, b), ___ptr_bitmap_128(a) < ___ptr_bitmap_128(b))
+      XCTAssertEqual(___ptr_comp_multi(b, a), ___ptr_bitmap_128(b) < ___ptr_bitmap_128(a))
+    }
+  }
 }
 
 final class NodeFlagTests0_1_2_3_4_5_6: TreeFixture0_1_2_3_4_5_6 {
@@ -59,6 +67,14 @@ final class NodeFlagTests0_1_2_3_4_5_6: TreeFixture0_1_2_3_4_5_6 {
                      uintCenteredToInt(___ptr_bitmap(a)))
       XCTAssertEqual(___ptr_comp_multi(a, b), ___ptr_comp_bitmap(a, b))
       XCTAssertEqual(___ptr_comp_multi(b, a), ___ptr_comp_bitmap(b, a))
+    }
+  }
+  
+  func testNodeFlag128() {
+    XCTAssertEqual(___ptr_bitmap_128(__root()), 1 << (UInt128.bitWidth - 1))
+    for (a,b) in (0...6).permutations(ofCount: 2).map({ ($0[0],$0[1])}) {
+      XCTAssertEqual(___ptr_comp_multi(a, b), ___ptr_bitmap_128(a) < ___ptr_bitmap_128(b))
+      XCTAssertEqual(___ptr_comp_multi(b, a), ___ptr_bitmap_128(b) < ___ptr_bitmap_128(a))
     }
   }
 }