+++ Tests/RedBlackTreeTests/tree/_NodeRef.swift
@@ -4,6 +4,7 @@ import Foundation
 
 extension _NodeRef {
     var index: Int! {
+#if false
         switch self {
         case .__right_(let p):
             return p
@@ -12,6 +13,14 @@ extension _NodeRef {
         default:
             return nil
         }
+#else
+      switch self {
+      case ~0:
+        return .end
+      default:
+        return .init(bitPattern: self & ~(1 << (Self.bitWidth - 1)))
+      }
+#endif
     }
 }
 #endif