+++ CHANGELOG.md
@@ -5,12 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2025-?-?
+### Added
+- サブシーケンスのサブスクリプトにRangeExpression対応を追加
+### Fixed
+- DictionaryとMapのinsertのバグを修正
+
 ## [0.2.0] - 2025-10-12
 ### Added
+- RedBlackTreeMapを追加
 - RedBlackTreeIndexを追加
 - RedBlackTreeSliceを追加
 - RangeExpression対応を追加
 ### Changed
+- RedBlackTreeMultiMapのKeyValueをタプルから構造体に変更
 - IndexをRedBlackTreeIndexに変更
 - SubSequenceをRedBlackTreeSliceに変更
 - 内部Indexの比較アルゴリズムを変更